+++ package/thunderstore/CheatDeath/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.3
+* fix for NRE on startup
+* status effect icon could be replaced by putting CheatDeath.png file next to plugin dll
+* new options to set custom text on fatal blow both server synced and client specific
+
 # 1.0.2
 * fix for consecutive effect proc not applying protection effects